Refactor directory and file
