+++ package.json
@@ -2,11 +2,11 @@
   "name": "example1",
   "version": "1.0.0",
   "description": "mediasoup tutorials",
-  "main": "app.js",
+  "main": "src/app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon app.js",
-    "watch": "watchify public/index.js -o public/bundle.js -v"
+    "start": "nodemon src/app.js",
+    "watch": "watchify src/public/index.js -o src/public/bundle.js -v"
   },
   "author": "amir eshaq",
   "license": "ISC",
